Renamed module -> spice_module for consistency
